requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
curl_cffi==0.7.1
ddddocr==1.5.5
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.8
loguru==0.7.2
mpmath==1.3.0
numpy==2.1.1
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==5.28.0
pycparser==2.22
pyreadline3==3.4.1
python-dotenv==1.0.1
requests==2.32.3
sympy==1.13.2
urllib3==2.2.2
win32-setctime==1.1.0
